backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.30
asyncpg==0.29.0
redis==5.0.4
greenlet==3.0.3
openai==1.30.1
PyJWT==2.9.0
python-multipart==0.0.9
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
pydantic-settings==2.2.1
pycrdt>=0.12.13,<0.13.0
aiosqlite==0.20.0
pytest==8.3.2
pytest-asyncio==0.24.0
